embeddings/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

